Move tests to prepare for refactor to move covered methods
