+++ packages/framework/tests/Feature/YamlConfigurationServiceTest.php
@@ -15,32 +15,5 @@ use Illuminate\Support\Facades\Config;
  */
 class YamlConfigurationServiceTest extends TestCase
 {
-    public function test_changes_in_yaml_file_override_changes_in_site_config()
-    {
-        $this->assertEquals('HydePHP', Config::get('site.name'));
-        $this->file('hyde.yml', 'name: Foo');
-        YamlConfigurationService::boot();
-        $this->assertEquals('Foo', Config::get('site.name'));
-    }
-
-    public function test_changes_in_yaml_file_override_changes_in_site_config_when_using_yaml_extension()
-    {
-        $this->assertEquals('HydePHP', Config::get('site.name'));
-        $this->file('hyde.yaml', 'name: Foo');
-        YamlConfigurationService::boot();
-        $this->assertEquals('Foo', Config::get('site.name'));
-    }
-
-    public function test_service_gracefully_handles_missing_file()
-    {
-        YamlConfigurationService::boot();
-        $this->assertEquals('HydePHP', Config::get('site.name'));
-    }
-
-    public function test_service_gracefully_handles_empty_file()
-    {
-        $this->file('hyde.yml', '');
-        YamlConfigurationService::boot();
-        $this->assertEquals('HydePHP', Config::get('site.name'));
-    }
+    //
 }
